Move const from utils/.
